monitoring/requirements.txt
fastapi==0.108.0
uvicorn==0.25.0
httpx==0.26.0
python-dotenv==1.0.0
supabase==2.3.1
apscheduler==3.10.4
pydantic==2.5.3
python-multipart==0.0.6
ujson==5.9.0
uuid==1.30
python-jose==3.3.0
cryptography==41.0.7
psycopg2-binary==2.9.9
redis==5.0.1 